Move search logic from IndexPage and app.cache.searched into SearchState
