+++ js/src/forum/components/HeaderSecondary.js
@@ -34,7 +34,7 @@ export default class HeaderSecondary extends Component {
   items() {
     const items = new ItemList();
 
-    items.add('search', Search.component(), 30);
+    items.add('search', Search.component({ state: app.search }), 30);
 
     if (app.forum.attribute('showLanguageSelector') && Object.keys(app.data.locales).length > 1) {
       const locales = [];
